Change all relative URLs of images to absolute URLs
It seems that the relative URLs for our images in assets/ that used to work no longer work. Looking online, it seems that GitHub is actually supposed to support these URLs. But maybe they have a regression and they're not supporting them here?

In any case, out of expediency I've converted all image URLs to absolute URLs.